Added infrastructure to allow multi-file assembly and external subroutine calls. I've placed a tutorial for how this works at the start of the UserInput file. Unfortunately, the subroutines aren't cooperating right now. I can't seem to find any words at all in the word list. Could we perchance change this to a comma or space delineated list?
I've marked out the blocks where logic from outside the UserInput shell needs to happen. Please have a look if you get time and feel free to insert the proper way of running those operations so I can understand your subroutines better.

I am also receiving constant errors with how the backend jumbles the input words. It seems it's treating the newline character as one of the characters to jumble. For instance, one time I got the following as the Letters array:
ITY
ARBAR

It looks like what happened is it wrap-shifted the word BARBARITY left 6 characters, but went and wrote the newline character in as character 4 in the array. Honestly, I don't even need the newline at the end of the array, since it never prints. I just need the first 9 slots of the array to be the characters in question.

Small changes to all files. Added the .globl marker. Feel free to add more .globl labels if need be.