BAREBONES USERINPUT ALPHA
Finally figured out how to get backend to generate word list and had to modify it to store ra to stack. Backed is more or less integrated. It's not pretty, but it's doing what it's supposed to and performing as expected. StringCheck and Scoring are broken. Problems include:

1) It does not work as indicated. I have no clue what number it's passing back, but it's not the length of the string. It even passes back a high number when I enter incorrect words.
2) It adds 4 seconds to timer for both failures and matches (Daniel indicated he intended for it to add, just not always.)
3) CheckString calls Scoring, eating up a save and recovery from the stack, for two lines of code, and then never actually passes back the score. Daniel indicated he would just fuse the Scoring logic into CheckString. For your reference, UserInput is using s3 for the score, so it can be added directly.
4) There is no logic to track which words have been used.

Since UserInput has a barebones implementation that should make one single game playable, I'm uploading this as is so Daniel can get his subroutines working with it. I'll be back later to start working on a cleaner implementation. Everyone should more or less start tinkering with ways that we can nab extra credit.

Julian's Checklist of Things To Complete UserInput: Make games restart, make it so grid changes letters (they stay A-F right now), play with writing them to MMIO Keyboard simulator, modify timer logic to adjust its speed