This is about as polished as it gets. There is still one major bug, though. See below.
Updates:

The scoring system has been updated to show the total found of the total possible.
The game ends on three incorrect guesses or if all words are found.
The timer modifies how long the program waits to adjust to different speed computers.
HUD shows number of tries remaining.
StringCheck was altered in one line to reflect new scoring method.
Backend was altered on one line to reflect new scoring method. (NEEDS DOUBLE-CHECKING!)

Issues:
I seem to be getting the wrong number of total words from the backend. I've tagged the line I added with JULIAN and it's the only line that sets s4 in that file. Can one of the guys who knows the backend better adjust this if need be so that s4 is getting the right number of total words in the word list?